+++ tket/src/Transformations/include/Transformations/MeasurePass.hpp
@@ -26,6 +26,33 @@ namespace Transforms {
  */
 Transform delay_measures();
 
+namespace DelayMeasures {
+
+/** Commute all measurement gates to the end of the circuit.
+ * @param circ The circuit to delay measurements in.
+ * @param dry_run If true, do not modify the circuit, just check if it is
+ * possible to delay.
+ * @throws CircuitInvalidity if it is not possible to delay and dry_run is
+ * false.
+ * @return A pair of booleans. The first indicates when the circuit was changed,
+ * and the second indicates if there where no undelayable gates (i.e. there
+ * where no errors).
+ **/
+std::pair<bool, bool> run_delay_measures(Circuit& circ, bool dry_run = false);
+
+/**
+ * Gathers all end-measurements, and adds the measured units to the list.
+ * Rejects gates acting on measured_units, and terminates early.
+ * Applies recursively for CircBoxes and Conditionals.
+ * @param com The command to check.
+ * @param measured_units The list of measured units to add to, initialy
+ * populated with previously-measured units.
+ * @return Whether there are no mid-circuit measurements.
+ **/
+bool check_only_end_measures(const Command& com, unit_set_t& measured_units);
+
+}  // namespace DelayMeasures
+
 }  // namespace Transforms
 
 }  // namespace tket